requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

PyPDF2==3.0.1
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.27.4
huggingface_hub==0.14.1
llama-cpp-python==0.2.60
python-dotenv==1.0.1
langchain==0.2.0
